refactor(app): extract bottom bar + nav panels from ReaderActivity
Behavior-preserving extraction (#115), second slice, same controller +
Host pattern: BottomBarController owns the KOReader-style bottom bar
(page slider, chapter jumps, control row), go-to-page entry, bookmark
toggle + list, Contents (TOC), the handwritten-annotations list, the
library picker, and go-home. Chapter-label helpers move with their only
caller. Sub-surfaces owned by sibling controllers (Search/Export/Dicts/
Adjust) open through Host callbacks; page jumps, zoom, and the palm
guard stay in the shell.

ReaderActivity drops from 2,954 to 2,550 lines.
